requirements.txt
langchain
langgraph
langchain_aws
langsmith
python-dotenv
langgraph-cli[inmem]
pandas